Added prototype ASN.1 encoding of pubkey files, added makefile
Added encoding of public key files via asn1c (needs to be installed), not integrated into keygen or rsa yet.
Also added Makefile which can be used with 'make' to build the entire project and 'make clean' to clean up. ONLY use make to compile this project from now on.